% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rlg_c1}
\alias{rlg_c1}
\title{Internal function for concentration steps (initializations) in rlg}
\usage{
rlg_c1(x, d, alpha = 0.05, niter1 = 3L)
}
\arguments{
\item{x}{Rcpp::NumericMatrix, The input data.}

\item{d}{Rcpp::NumericVector, A numeric vector of length equal to the number 
of clusters to be detected. 
Each component of vector \code{d} indicates the intrinsic dimension of the affine subspace 
where observations on that cluster are going to be clustered. All the elements 
of vector \code{d} should be smaller than the problem dimension minus 1.}

\item{alpha}{double, The proportion of observations to be trimmed.}

\item{niter1}{int, The number of concentration steps to be performed for the 
nstart initializations.}
}
\description{
Internal function for concentration steps (initializations) in rlg
}
