% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tclust_c2}
\alias{tclust_c2}
\title{Internal function for concentration steps (refinement) in tclust2}
\usage{
tclust_c2(
  x,
  k,
  cluster,
  alpha = 0.05,
  restrC = 0L,
  deterC = FALSE,
  restr_fact = 12,
  cshape = 1e+10,
  niter2 = 20L,
  opt = "HARD",
  equal_weights = FALSE,
  zero_tol = 1e-16
)
}
\arguments{
\item{x}{Rcpp::NumericMatrix, The input data.}

\item{k}{The number of clusters initially searched for.}

\item{cluster}{arma::uvec A numerical vector of size \code{n} containing the 
cluster assignment for each observation. Cluster names are integer numbers 
from 1 to k, 0 indicates trimmed observations. Note that it could be empty 
clusters with no observations when \code{equal_weights=FALSE}.}

\item{alpha}{double, The proportion of observations to be trimmed.}

\item{restrC}{Restriction type (0 for restriction on eigenvalues or determinant)}

\item{deterC}{Restriction on determinants (true or false)}

\item{restr_fact}{The constant \code{restr_fact >= 1} constrains the allowed 
differences among group scatters in terms of eigenvalues ratio. Larger values 
imply larger differences of group scatters, a value of 1 specifies the 
strongest restriction.}

\item{cshape}{Shape constraint}

\item{niter2}{The maximum number of concentration steps to be performed for the 
\code{nkeep} solutions kept for further iteration. The concentration steps are 
stopped, whenever two consecutive steps lead to the same data partition.}

\item{opt}{Define the target function to be optimized. A classification likelihood 
target function is considered if \code{opt="HARD"} and a mixture classification 
likelihood if \code{opt="MIXT"}.}

\item{equal_weights}{A logical value, specifying whether equal cluster weights 
shall be considered in the concentration and assignment steps.}

\item{zero_tol}{The zero tolerance used. By default set to 1e-16.}
}
\description{
Internal function for concentration steps (refinement) in tclust2
}
