% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tkmeans_c1}
\alias{tkmeans_c1}
\title{Internal function for concentration steps (initializations) in tkmeans2}
\usage{
tkmeans_c1(x, k, alpha = 0.05, niter1 = 3L, zero_tol = 1e-16, points = NULL)
}
\arguments{
\item{x}{Rcpp::NumericMatrix, The input data.}

\item{k}{The number of clusters initially searched for.}

\item{alpha}{double, The proportion of observations to be trimmed.}

\item{niter1}{int, The number of concentration steps to be performed for the 
nstart initializations.}

\item{zero_tol}{The zero tolerance used. By default set to 1e-16.}

\item{points}{Optional initial mean vectors, \code{NULL} or a matrix with \code{k} 
vectors used as means to initialize the algorithm. If initial mean vectors are 
specified, \code{nstart} should be 1 (otherwise the same initial means are 
used for all runs).}
}
\description{
Internal function for concentration steps (initializations) in tkmeans2
}
