% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tkmeans_c2}
\alias{tkmeans_c2}
\title{Internal function for concentration steps (refinement) in tkmeans2}
\usage{
tkmeans_c2(x, k, cluster, alpha = 0.05, niter2 = 20L, zero_tol = 1e-16)
}
\arguments{
\item{x}{Rcpp::NumericMatrix, The input data.}

\item{k}{The number of clusters initially searched for.}

\item{cluster}{arma::uvec A numerical vector of size \code{n} containing the 
cluster assignment for each observation. Cluster names are integer numbers 
from 1 to k, 0 indicates trimmed observations.}

\item{alpha}{double, The proportion of observations to be trimmed.}

\item{niter2}{The maximum number of concentration steps to be performed for the 
\code{nkeep} solutions kept for further iteration. The concentration steps are 
stopped, whenever two consecutive steps lead to the same data partition.}

\item{zero_tol}{The zero tolerance used. By default set to 1e-16.}
}
\description{
Internal function for concentration steps (refinement) in tkmeans2
}
